合并

+++ Server/xiaomi_mall/src/main/java/com/xiaomi_mall/controller/BackProductController.java
@@ -1,5 +1,7 @@
 package com.xiaomi_mall.controller;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.xiaomi_mall.config.Result;
 import com.xiaomi_mall.dto.*;
 import com.xiaomi_mall.enity.Category;
@@ -154,16 +156,30 @@ public class BackProductController {
 
     @PreAuthorize("hasAnyAuthority('普通管理员', '超级管理员')")
     @ApiOperation("删除Attribute接口")
-    @DeleteMapping("/deleteAttribute")
+    @DeleteMapping("/deleteAttribute/{attributeId}")
     public Result deleteAttribute(@PathVariable Integer attributeId) {
-        skuAttributeService.removeById(attributeId);
-        List<Integer> attributeValueId = new ArrayList<>();
-        for (SkuAttributeValue value: skuAttributeValueService.list()) {
-            if (value.getAttributeId() == attributeId)
-                attributeValueId.add(value.getValueId());
+        try
+        {
+            LambdaQueryWrapper<SkuAttributeValue> skuAttributeValueQueryWrapper = new LambdaQueryWrapper<>();
+            skuAttributeValueQueryWrapper.eq(SkuAttributeValue::getAttributeId, attributeId);
+            int cnt = skuAttributeValueService.count(skuAttributeValueQueryWrapper);
+            if(cnt != 0)
+                return Result.errorResult2(924, "该规格还有值哦，不能删除~");
+
+            skuAttributeService.removeById(attributeId);
+            List<Integer> attributeValueId = new ArrayList<>();
+            for (SkuAttributeValue value: skuAttributeValueService.list()) {
+                if (value.getAttributeId() == attributeId)
+                    attributeValueId.add(value.getValueId());
+            }
+            skuAttributeValueService.removeByIds(attributeValueId);
+            return Result.okResult();
         }
-        skuAttributeValueService.removeByIds(attributeValueId);
-        return Result.okResult();
+        catch (Exception e)
+        {
+            return Result.errorResult2(500, e.getMessage());
+        }
+
     }
 
     @PreAuthorize("hasAnyAuthority('普通管理员', '超级管理员')")
